refactor: extract screen isolation into separate modules to meet 1000-line limit
Both isolation.js (1016 lines) and isolation.rs (1048 lines) exceeded the
1000-line file size limit enforced by CI.

- Extract screen version detection and log capture logic from isolation.js
  into new js/src/lib/screen-isolation.js module
- Extract screen-related functions from isolation.rs into new
  rust/src/lib/isolation_screen.rs module
- Add rust/changelog.d/97.md changelog fragment for issue #96 fix

After refactoring:
- isolation.js: 758 lines (was 1016)
- isolation.rs: 851 lines (was 1048)
- screen-isolation.js: 309 lines (new)
- isolation_screen.rs: 215 lines (new)
