+++ rust/changelog.d/97.md
@@ -1,0 +1,19 @@
+fix: capture output from quick-completing commands in screen isolation (issue #96)
+
+When running a short-lived command like `agent --version` through screen isolation,
+the output was silently lost because GNU Screen's internal log buffer flushes every
+10 seconds by default. For commands that complete faster than this, the buffer may
+not be flushed to the log file before the screen session terminates.
+
+**Fix:** A temporary screenrc file with `logfile flush 0` is passed to screen via
+the `-c` option. This forces screen to flush the log buffer after every write,
+eliminating the 10-second flush delay for quick-completing commands.
+
+A retry mechanism is also added for the tee fallback path (older screen < 4.5.1)
+to handle the TOCTOU race where the log file appears empty when first read
+immediately after session completion.
+
+The screen-related functions have also been extracted from `isolation.rs` into a
+new `isolation_screen.rs` module to keep file sizes under the 1000-line limit.
+
+Fixes #96
